feat: lsp e2e test. Move unit tests to implement file and add e2e test frame to lsp/tests.rs
