+++ kclvm/tools/src/LSP/src/find_ref/tests.rs
@@ -32,11 +32,9 @@ mod tests {
 
     #[test]
     fn test_line_to_words() {
-        let datas = vec![
-            "alice_first_name = \"alice\"",
+        let datas = ["alice_first_name = \"alice\"",
             "0lice_first_name = \"alic0\"",
-            "alice = p.Parent { name: \"alice\" }",
-        ];
+            "alice = p.Parent { name: \"alice\" }"];
         let expect = vec![
             vec![
                 LineWord {
