+++ kclvm/tools/src/LSP/src/lib.rs
@@ -3,16 +3,18 @@ mod completion;
 mod config;
 mod db;
 mod dispatcher;
+mod document_symbol;
 mod find_ref;
 mod formatting;
 mod from_lsp;
-mod notification;
-mod state;
-mod to_lsp;
-mod util;
-
-mod document_symbol;
 mod goto_def;
 mod hover;
+mod main_loop;
+mod notification;
 mod quick_fix;
 mod request;
+mod state;
+#[cfg(test)]
+mod tests;
+mod to_lsp;
+mod util;
