+++ kclvm/tools/src/LSP/src/test_data/diagnostics.k
@@ -7,4 +7,3 @@ c: Person = Person {
 
 d = 1
 d = 2
-
